requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.2
Flask-Limiter==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
python-dotenv==1.1.1
rich==14.1.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
